UPDATE: refactor auth route
