+++ client/src/app/layout.tsx
@@ -1,5 +1,8 @@
 import "./globals.css";
-import "material-symbols";
+import "@/assets/css/goongjs.css";
+import "material-symbols/outlined.css";
+import "react-toastify/dist/ReactToastify.css";
+
 import type { Metadata } from "next";
 import { Lexend } from "next/font/google";
 import { ToastContainer } from "react-toastify";
@@ -14,11 +17,6 @@ export const metadata: Metadata = {
 export default function RootLayout({ children, }: { children: React.ReactNode; }) {
   return (
     <html lang="en">
-      <head>
-        <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Rounded:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-        <link href='https://cdn.jsdelivr.net/npm/@goongmaps/goong-js@1.0.9/dist/goong-js.css' rel='stylesheet' />
-
-      </head>
       <body className={inter.className}>
         <ToastContainer
           limit={1}
@@ -34,6 +32,7 @@ export default function RootLayout({ children, }: { children: React.ReactNode; }
           theme="light"
           className={"text-[14px]"}
         />
+        
         {children}
       </body>
     </html>
