--- client/src/components/AuthContainer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AuthContainer";
