+++ client/src/components/FormWrapper/FormWrapper.tsx
@@ -1,25 +1,31 @@
 import Logo from "@/icons/Logo";
 import { faGithub } from "@fortawesome/free-brands-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import React from "react";
+
 type Props = {
-  children: React.ReactNode;
+  children: React.ReactNode
 };
+
 const FormWrapper = ({ children }: Props) => {
   return (
     <div className="h-full flex flex-col select-none">
       <header className="flex flex-col mb-[44px]">
         <div className="flex flex-row items-center">
-          <Logo width="60" height="30" className={"text-[#FDB813]"}></Logo>
-          <h2 className="text-[16px] text-black ml-2 font-[700]">
+          <Logo width="60" height="30" className={"text-cyellow-500"}></Logo>
+          <h2 className="text-[16px] text-cgray-600 ml-2 font-[700]">
             Hệ thống quản lý chuyển phát
           </h2>
         </div>
       </header>
+
       {children}
+
       <footer className="mt-auto mb-4 flex flex-row items-center gap-4 justify-start text-[12px] text-gray-400 font-medium select-none ">
         <span className="hover:cursor-pointer hover:opacity-60 active:opacity-80">
           @devbygroup6
         </span>
+
         <span className="flex flex-row items-center hover:cursor-pointer hover:opacity-60 active:opacity-80">
           <FontAwesomeIcon
             icon={faGithub}
@@ -27,6 +33,7 @@ const FormWrapper = ({ children }: Props) => {
           />
           github
         </span>
+
         <span className="hover:cursor-pointer hover:opacity-60 active:opacity-80">
           <span className="text-[12px] font-semibold">SOS</span>
           <span className="ml-1">Liên hệ kỹ thuật</span>
@@ -35,4 +42,5 @@ const FormWrapper = ({ children }: Props) => {
     </div>
   );
 };
+
 export default FormWrapper;
