--- client/src/layouts/login/LoginContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./LoginContent";
